Change: minor changes
 - enable CORS for client
 - make account balance Double (from Long)
 - fix color reges (allow 0 character)
 - add user to athentication and register response

+++ src/main/java/com/babakjan/moneybag/entity/Account.java
@@ -33,7 +33,7 @@ public class Account {
     private String currency;
 
     @Nonnull
-    private Long balance;
+    private Double balance;
 
     @Nonnull
     private String color = "#6290ff";
